- Fixed members of union group_64 to have correct length for 16 and 32 bit fields

+++ include/scandal/tritium.h
@@ -30,8 +30,8 @@
 typedef union _group_64 {
 	float data_fp[2];
 	unsigned char data_u8[8];
-	unsigned int data_u16[4];
-	unsigned long data_u32[2];
+	unsigned short int data_u16[4];
+	unsigned int data_u32[2];
 } group_64;
 
 typedef union _group_32 {
